refactor: v2 컨벤션에 맞게 크롤링 로직 위치 이동
