+++ app-main/src/main/java/net/causw/app/main/domain/integration/crawled/service/CrawledToPostTransferService.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.integration.crawled.service.v1;
+package net.causw.app.main.domain.integration.crawled.service;
 
 import java.util.ArrayList;
 import java.util.List;
