+++ app-main/src/test/java/net/causw/app/main/domain/integration/crawled/service/CrawledNoticeFixture.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.integration.crawled.service.v1;
+package net.causw.app.main.domain.integration.crawled.service;
 
 import java.time.LocalDate;
 import java.util.Collections;
