+++ app-main/src/test/java/net/causw/app/main/domain/integration/crawled/service/CrawledToPostTransferServiceTest.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.integration.crawled.service.v1;
+package net.causw.app.main.domain.integration.crawled.service;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
