+++ app-main/src/test/java/net/causw/app/main/domain/integration/crawled/service/implementation/JsoupCrawlerTest.java
@@ -1,4 +1,4 @@
-package net.causw.app.main.domain.integration.crawled.service.v1;
+package net.causw.app.main.domain.integration.crawled.service.implementation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
